Refactor to MetadataParserTrait
